ICMSLST-507 Move common management features from firearms to _import
The features covered by this PR are common to all the import
applications.

Move ILB Admin case management to:
 - Take/Release Ownership of a case
 - Manage the case
 - View the case

Move Importer case management to:
 - Withdraw a case
 - Retract a withdrawal
 - View a case
