+++ web/domains/case/_import/firearms/models.py
@@ -3,7 +3,6 @@ from django.db import models
 from web.domains.constabulary.models import Constabulary
 from web.domains.file.models import File
 from web.domains.firearms.models import FirearmsAuthority
-from web.domains.user.models import User
 
 from ..models import ImportApplication
 
@@ -72,38 +71,6 @@ class VerifiedCertificate(models.Model):
     )
     firearms_authority = models.ForeignKey(
         FirearmsAuthority, on_delete=models.PROTECT, related_name="+"
-    )
-
-    created_datetime = models.DateTimeField(auto_now_add=True)
-    updated_datetime = models.DateTimeField(auto_now=True)
-
-
-class WithdrawImportApplication(models.Model):
-    STATUS_OPEN = "open"
-    STATUSES = (
-        (STATUS_OPEN, "OPEN"),
-        ("rejected", "REJECTED"),
-        ("accepted", "ACCEPTED"),
-    )
-    import_application = models.ForeignKey(
-        ImportApplication, on_delete=models.PROTECT, related_name="withdrawals"
-    )
-    is_active = models.BooleanField(default=True)
-    status = models.CharField(max_length=10, choices=STATUSES, default=STATUS_OPEN)
-    reason = models.TextField()
-    request_by = models.ForeignKey(
-        User,
-        on_delete=models.PROTECT,
-        related_name="+",
-    )
-
-    response = models.TextField()
-    response_by = models.ForeignKey(
-        User,
-        on_delete=models.PROTECT,
-        blank=True,
-        null=True,
-        related_name="+",
     )
 
     created_datetime = models.DateTimeField(auto_now_add=True)
